refactor(test-types,test-specs,test-execute): Deduplicate implicit gas-limit calculation
`calculate_max_transaction_gas_limit()` existed as two verbatim copies
on either side of the spec/execute split: one on `BlockchainTest` for
the fill path and one on `BaseExecute` for the execute formats. The
function encodes the policy for sizing implicit transaction gas limits
(even split of remaining environment gas, capping, state gas reservoir
handling), and both copies must agree for filled and executed tests to
behave the same: the recently added no-remaining-gas check already had
to be applied twice.

Hoist it to a single module-level function in
`test_types/transaction_types.py`, next to `Transaction.set_gas_limit()`,
so that all implicit gas-limit policy lives in one place. Neither class
used `self` or `cls` (both copies were static), so a free function is
the natural shape.

The parameter is narrowed from `env: Environment` to
`env_gas_limit: int`: the function only ever read `env.gas_limit`, and
taking the full `Environment` would have required `transaction_types`
to depend on `block_types` for a single field.
